Get Gerrit URL from manifest repository
Currently, `repoload` requires an environment variable called
`GERRIT_URL` to be present, which must contain the URL of the Gerrit
server. This behaviour is inflexible and requires a special environment
which has to be set up manually, although the information is available
in the context where `repoload` is used.

This patch makes the `GERRIT_URL` environment variable optional. If it
is not present, it will search for a `.repo/manifestes.git/config` file
in the current directory and all parent directories. Alternatively, if
the `ANDROID_BUILD_TOP` environment variable is set, it will use the
content of that variable as the directory containing the path.
The Gerrit URL is the `url` key of the `remote "origin"` configuration
section.

Signed-off-by: Florian Limberger <flo@purplekraken.com>